Error Detection Modes
Two of the simple error detection modes are examined in assembly language.  The first on is called CRC Cyclic Redundancy Check, and the other is Parity Bit Method. They are both applied on atmega128 MCU on a Proteus simulation. More of the detailed information can be found in description files.